Atualização da estrutura de pastas dos testes e2e
