+++ cypress/e2e/biblioteca/bibliotecaCriarAtividade.cy.js
@@ -1,13 +1,12 @@
 /// reference types="cypress" />
 import { fakerPT_BR } from '@faker-js/faker'
-import { getAuthToken } from '../support/authHelper' 
-import estruturaAtividades from '../support/pageObjects/estruturaAtividades'
-import formAtividades from '../support/pageObjects/formAtividades'
+import estruturaAtividades from '../../support/pageObjects/estruturaAtividades'
+import formAtividades from '../../support/pageObjects/formAtividades'
 
 describe('Criar atividade', () => {
     const esperaExplicita = 5000
 
-    let nomeConteudo, tipoConteudo, nomeAtividade
+    let nomeConteudo, tipoConteudo, nomeAtividade, listaConteudos
 
     let atividadeDefault = 'Novo 1'
 
@@ -42,8 +41,8 @@ describe('Criar atividade', () => {
         peso: 1,
         liberado: false,
         tipoAtividade: 'Vídeo',
-        enviarVideo: '',
-        descricaoArquivoVideo: {
+        enviarArquivo: '',
+        descricaoArquivo: {
             nome: '',
             tamanho: ''
         },
@@ -109,8 +108,8 @@ describe('Criar atividade', () => {
         peso: 1,
         liberado: false,
         tipoAtividade: 'Scorm',
-        enviarScorm: '',
-        descricaoArquivoScorm: {
+        enviarArquivo: '',
+        descricaoArquivo: {
             nome: '',
             tamanho: ''
         },
@@ -130,31 +129,27 @@ describe('Criar atividade', () => {
 
     beforeEach(() => {
         // Define o tipo de conteúdo
-        tipoConteudo = 'catalogo'
+        tipoConteudo = 'biblioteca'
 
         // Gera um nome aleatório para o conteúdo e para a atividade
         nomeConteudo = fakerPT_BR.commerce.productName()
         nomeAtividade = fakerPT_BR.commerce.productName()
 
-        // Obtém o token de autenticação 
-        getAuthToken()
-
-        // Exclui todos os catálogos antes de iniciar o teste
-        cy.excluirCatalogoViaApi()
-
-        // Cria um catálogo default
-        const body = {
-            name: nomeConteudo,
-            description: fakerPT_BR.lorem.sentence(5)
-        }
-        cy.criarCatalogoViaApi(body)
+        // Acessa página de biblioteca e gera uma lista com os conteúdos para serem excluídos
+        cy.acessarPgBiblioteca()
+        
+        listaConteudos = []
+        cy.listaConteudo(tipoConteudo, listaConteudos)
+        cy.excluirConteudo(null, tipoConteudo, listaConteudos)
+        
+        // Cria uma biblioteca
+        cy.criarBibliotecaDefault(nomeConteudo)    
     })
 
     it('1. Criar uma atividade default', () => {
         // CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -172,7 +167,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -195,12 +189,8 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             descricaoTexto: fakerPT_BR.lorem.sentence(10),
-            resumoAtividade: fakerPT_BR.lorem.sentence(5),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:08'
+            resumoAtividade: fakerPT_BR.lorem.sentence(5)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -232,7 +222,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -257,17 +246,12 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             enviarPdf: 'teste_pdf.pdf',
             descricaoArquivoPdf: {
                 nome: 'teste_pdf.pdf',
                 tamanho: '28102'
             },
-            seguranca: 'Somente Baixar',
-            resumoAtividade: fakerPT_BR.lorem.sentence(15),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:12'
+            resumoAtividade: fakerPT_BR.lorem.sentence(15)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -299,7 +283,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -324,19 +307,12 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             enviarVideo: 'teste_video.mp4',
             descricaoArquivoVideo: {
                 nome: 'teste_video.mp4',
                 tamanho: '50809927'
             },
-            marcarConcluidoVideo: true,
-            naoMostrarProgresso: true,
-            seguranca: 'Somente Baixar',
-            resumoAtividade: fakerPT_BR.lorem.sentence(50),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:25'
+            resumoAtividade: fakerPT_BR.lorem.sentence(50)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -368,7 +344,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -393,19 +368,13 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             youtube: true,
             vimeo: false,
             eventials: false,
             videoUrl: 'https://www.youtube.com/watch?v=OyTN-MF-OEg',
             marcarConcluidoVideoExterno: true,
             naoMostrarProgressoVideoExterno: true,
-            chatTwygo: true,
-            desabilitarChatFimTransmissao: true,
-            resumoAtividade: fakerPT_BR.lorem.sentence(8),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:02'
+            resumoAtividade: fakerPT_BR.lorem.sentence(8)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -437,7 +406,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -462,17 +430,12 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             enviarArquivo: 'Sophia_estudiosa.png',
             descricaoArquivo: {
                 nome: 'Sophia_estudiosa.png',
                 tamanho: '34264'
             },
-            seguranca: 'Somente Baixar',
-            resumoAtividade: fakerPT_BR.lorem.sentence(22),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:01'
+            resumoAtividade: fakerPT_BR.lorem.sentence(22)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -501,14 +464,13 @@ describe('Criar atividade', () => {
             selecionarQuestionario: fakerPT_BR.lorem.sentence(3)
         }
 
-        // CREATE
-        cy.log('## CREATE ##')
-
         // Criar questionário
         cy.criarQuestionarioDefault(dados.selecionarQuestionario)
 
-        // Acessar catálogo
-        cy.acessarPgCatalogo()
+        // CREATE
+        cy.log('## CREATE ##')
+
+        cy.acessarPgBiblioteca()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -535,19 +497,7 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
-            exibicaoPerguntas: 'Exibir perguntas diferentes a cada tentativa',
-            visualizacaoRespostas: 'Exibir Respondidas',
-            pontuacaoMinima: '50',
-            tentativas: '2',
-            percPontuacaoFinal: '70',
-            perguntasCat1: 'Todas',
-            perguntasCat2: 'Todas',
-            quantidadePerguntas: '',
-            resumoAtividade: fakerPT_BR.lorem.sentence(6),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '01:00'
+            resumoAtividade: fakerPT_BR.lorem.sentence(6)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
@@ -584,7 +534,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -612,14 +561,11 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             enviarScorm: 'teste_scorm2.zip',
             descricaoArquivoScorm: {
                 nome: 'teste_scorm2.zip',
                 tamanho: '8,3 MB'
             },    
-            marcarConcluidoScorm: true,
             resumoAtividade: fakerPT_BR.lorem.sentence(19)
         }
 
@@ -653,7 +599,6 @@ describe('Criar atividade', () => {
         // CREATE
         cy.log('## CREATE ##')
 
-        cy.acessarPgCatalogo()
         cy.addAtividadeConteudo(nomeConteudo, tipoConteudo)
         estruturaAtividades.adicionarAtividade()
         cy.salvarAtividades()
@@ -678,12 +623,8 @@ describe('Criar atividade', () => {
 
         const dadosUpdate = {
             titulo: nomeAtividade,
-            peso: fakerPT_BR.number.int({min: 1, max: 9}),
-            liberado: true,
             codigoCompartilhamento: '<iframe src= "https://kahoot.it/challenge/0857294?challenge-id=502fec44-a2dc-4312-807a-65e1d9bc4a4d_1695673333050" width=620 height=280></iframe>',
-            resumoAtividade: fakerPT_BR.lorem.sentence(2),
-            tempoMinPermanencia: true,
-            tempoMinPermanenciaValor: '00:03'
+            resumoAtividade: fakerPT_BR.lorem.sentence(2)
         }
 
         cy.preencherDadosAtividade(dadosUpdate, {limpar: true})
