+++ cypress/e2e/catalogo/catalogo.cy.js
@@ -1,13 +1,14 @@
 /// reference types="cypress" />
 import 'cypress-real-events/support'
 import { fakerPT_BR } from '@faker-js/faker'
-import { getAuthToken } from '../support/authHelper'
-import { gerarData } from '../support/utilsHelper'
+import { getAuthToken } from '../../support/authHelper'
+import { gerarData } from '../../support/utilsHelper'
+import formConteudos from "../../support/pageObjects/formConteudos"
 
-describe('trilha', () => {
-	let nome, tipoConteudo, categorias, novasCategorias, delCategorias, listaConteudos
+describe('catálogo', () => {
+	let nome, tipoConteudo, categorias, novasCategorias, delCategorias
 
-	// Campos e dados default do formulário de trilha
+	// Campos e dados default do formulário de catálogo
 	let formularioConteudo = {
 		nome: '',
 		dataInicio: '',
@@ -18,6 +19,7 @@ describe('trilha', () => {
 		tipo: 'Treinamento',
 		modalidade: 'Online',
 		sincronismo: 'Gravado',
+		canal: '',
 		cargaHoraria: '0',
 		vigencia: '0',
 		atualizarInscritos: false,
@@ -29,16 +31,31 @@ describe('trilha', () => {
 		estado: '',
 		pais: '',
 		emailResponsavel: Cypress.env('login'),
+		site: '',
 		notificarResponsavel: false,
+		rotuloContato: '',
+		hashtag: '',
 		addCategoria: '',
 		removerCategoria: '',
+		permiteAnexo: 'Desabilitado',
+		mensagemAnexo: '',
+		statusIframeAnexo: false,
+		visualizacao: 'Inscritos',
 		situacao: 'Em desenvolvimento',
-		exigeConfirmacao: 'Desabilitado'
+		notificarConcluirPrimeiraAula: 'Não',
+		notificarUsuarios: 'Não',
+		diasTeste: '0',
+		habilitarDiasTeste: false,
+		exigeConfirmacao: 'Desabilitado',
+		valorInscricao: '0,00',
+		habilitarPagamento: false,
+		nrParcelas: '1',
+		valorAcrescimo: '0.0',
 	}
 
 	beforeEach(() => {
 		// Define o tipo de conteúdo
-		tipoConteudo = 'trilha'
+		tipoConteudo = 'catalogo'
 
 		// Gera um nome aleatório para o conteúdo
 		nome = fakerPT_BR.commerce.productName()
@@ -47,37 +64,34 @@ describe('trilha', () => {
 		categorias = []
 		novasCategorias = []
 		delCategorias = []
-
-		// Exclui todos os cursos da lista de conteúdos
-		getAuthToken()
-		cy.excluirCursoViaApi()
-
-		// Exclui todos os conteúdos do tipo trilha antes de iniciar o teste
 		
-		listaConteudos = []
-		cy.listaConteudo(tipoConteudo, listaConteudos)
-		cy.excluirConteudo(null, tipoConteudo, listaConteudos)		
-	})
+		// Obtem o token de autenticação
+		getAuthToken()
 
-	it('1. CRUD trilha com dados default', () =>{
-		// Massa de dados para criação da trilha
-        const conteudo = {
-			nome: nome,
-            descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+		// Exclui todos os catálogos antes de iniciar o teste
+		cy.excluirCatalogoViaApi()
+	})
+	
+	it('1. CRUD catalogo com dados default', () => {
+		// Massa de dados para criação do catálogo
+		const conteudo = {
+			nome: fakerPT_BR.commerce.productName(),
+			descricao: fakerPT_BR.commerce.productDescription()
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
-        
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
+
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
@@ -92,10 +106,11 @@ describe('trilha', () => {
 			horaInicio: '12:00',
 			dataFim: '29/04/2024',
 			horaFim: '22:00',
-			descricao: `${fakerPT_BR.commerce.productDescription()} editado do evento ${novoNome}`,
+			descricao: `Descrição editada do curso nome: ${novoNome}`,
 			tipo: 'Congresso',
 			modalidade: 'Presencial',
 			sincronismo: 'Ao vivo',
+			canal: 'Outros',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			atualizarInscritos: true,
@@ -107,12 +122,27 @@ describe('trilha', () => {
 			estado: 'PR',
 			pais: 'Brasil',
 			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
 			notificarResponsavel: true,
+			rotuloContato: 'Contato',
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			mensagemAnexo: 'Insira o anexo do Catálogo do evento:',
+			visualizacao: 'Público',
 			situacao: 'Liberado',
-			exigeConfirmacao: 'Habilitado'
+			notificarConcluirPrimeiraAula: 'Sim',
+			notificarUsuarios: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: true,
+			exigeConfirmacao: 'Habilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
+			habilitarPagamento: true,
+			nrParcelas: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 })
 		}
-
+		
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
 		cy.salvarConteudo(conteudoEdit.nome, tipoConteudo)
 
@@ -121,9 +151,8 @@ describe('trilha', () => {
 
 		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
 
-		const todasCategorias = [...categorias, ...novasCategorias]
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
-		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// DELETE
 		cy.log('## DELETE ##')
@@ -132,8 +161,8 @@ describe('trilha', () => {
 		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
 	})
 
-	it('2. CRUD trilha liberada, com confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
+	it('2. CRUD catalogo liberado, com anexo, com pagamento, sem acréscimo, com confirmação, com visualização para inscritos', () => {
+		// Massa de dados para criação do catálogo
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
@@ -141,10 +170,11 @@ describe('trilha', () => {
 			horaInicio: '01:00',
 			dataFim: '29/04/2024',
 			horaFim: '23:00',
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
 			tipo: 'Congresso',
 			modalidade: 'Presencial',
 			sincronismo: 'Ao vivo',
+			canal: 'Outros',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			local: 'Centro de Eventos',
@@ -155,51 +185,78 @@ describe('trilha', () => {
 			estado: 'SC',
 			pais: 'Brasil',
 			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
 			notificarResponsavel: true,
+			rotuloContato: 'Fale conosco',
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			mensagemAnexo: `Insira o anexo do Catálogo do evento: ${nome}`,
+			visualizacao: 'Inscritos',
 			situacao: 'Liberado',
-			exigeConfirmacao: 'Habilitado'
+			notificarConcluirPrimeiraAula: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: true,
+			exigeConfirmacao: 'Habilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
+			habilitarPagamento: true,
+			nrParcelas: fakerPT_BR.number.int({ min: 1, max: 9 })
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-        cy.validarDadosConteudo(dadosParaValidar, categorias)
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// UPDATE
 		cy.log('## UPDATE ##')
 
 		const novoNome = fakerPT_BR.commerce.productName()
-		novasCategorias = [`Cat3-${fakerPT_BR.hacker.noun()}`, `Cat4-${fakerPT_BR.hacker.noun()}`]
+		categorias = [`Cat3-${fakerPT_BR.hacker.noun()}`]
 		const conteudoEdit = {
 			nome: novoNome,
 			dataInicio: '01/01/2023',
 			horaInicio: '00:01',
 			dataFim: '31/01/2025',
 			horaFim: '23:59',
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${novoNome}`,
+			descricao: `Descrição editada do curso nome: ${novoNome}`,
 			tipo: 'Feira',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
+			canal: '',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			atualizarInscritos: true,
 			local: 'T&D Connect',
 			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
 			notificarResponsavel: false,
-			addCategoria: novasCategorias,
+			rotuloContato: 'Contato',
+			hashtag: fakerPT_BR.hacker.adjective(),
+			addCategoria: categorias,
+			permiteAnexo: 'Desabilitado',
+			statusIframeAnexo: false,
+			visualizacao: 'Público',
 			situacao: 'Suspenso',
+			notificarConcluirPrimeiraAula: 'Não',
+			notificarUsuarios: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: false,
 			exigeConfirmacao: 'Desabilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
+			habilitarPagamento: false
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -208,7 +265,11 @@ describe('trilha', () => {
 		// READ-UPDATE
 		cy.log('## READ-UPDATE ##')
 
+		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
+
+		// Após salvar a alteração de modalidade de Presencial para Online, o campo de CEP não é limpo como os demais campos de endereço
 		let dadosEspecificos = {
+			cep: '85804-455',
 			endereco: '',
 			complemento: '',
 			cidade: '',
@@ -216,78 +277,104 @@ describe('trilha', () => {
 			pais: ''
 		}
 
-		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
-
-		const todasCategorias = [...categorias, ...novasCategorias]
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit, ...dadosEspecificos }		
-		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit, ...dadosEspecificos }
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// DELETE
 		cy.log('## DELETE ##')
-		
+
 		cy.cancelarFormularioConteudo(tipoConteudo)
 		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
 	})
 
-	it('3. CRUD trilha liberada, sem confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
-		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
+	it('3. CRUD catalogo liberado, com anexo, com pagamento, c/acréscimo, sem confirmação, com visualização para público', () => {
+		// Massa de dados para criação do catálogo
+		const categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
 			dataInicio: '29/01/2023',
 			horaInicio: '11:20',
 			dataFim: '29/01/2024',
 			horaFim: '22:02',
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
 			tipo: 'Feira',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
+			canal: 'Em companhia',
 			cargaHoraria: fakerPT_BR.number.int({ min: 10, max: 99 }),
 			vigencia: fakerPT_BR.number.int({ min: 10, max: 99 }),
 			local: 'Youtube',
 			emailResponsavel: fakerPT_BR.internet.email(),
-			notificarResponsavel: false,
+			site: fakerPT_BR.internet.url(),
+			rotuloContato: 'Entre em contato conosco para mais informações.',
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
-			situacao: 'Liberado'
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			visualizacao: 'Público',
+			situacao: 'Liberado',
+			notificarConcluirPrimeiraAula: 'Sim',
+			notificarUsuarios: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 10, max: 99 }),
+			habilitarDiasTeste: true,
+			valorInscricao: fakerPT_BR.commerce.price({ min: 50, max: 99 }),
+			habilitarPagamento: true,
+			nrParcelas: fakerPT_BR.number.int({ min: 10, max: 12 }),
+			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 })
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-        cy.validarDadosConteudo(dadosParaValidar, categorias)
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// UPDATE
 		cy.log('## UPDATE ##')
 
 		const novoNome = fakerPT_BR.commerce.productName()
-		novasCategorias = [`Cat4-${fakerPT_BR.hacker.noun()}`, `Cat5-${fakerPT_BR.hacker.noun()}`]		
+		novasCategorias = [`Cat4-${fakerPT_BR.hacker.noun()}`, `Cat5-${fakerPT_BR.hacker.noun()}`]
 		const conteudoEdit = {
 			nome: novoNome,
 			dataInicio: '10/03/2000',
 			horaInicio: '00:00',
 			dataFim: '31/12/2050',
 			horaFim: '03:40',
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${novoNome}`,
+			descricao: `Descrição editada do curso nome: ${novoNome}`,
 			tipo: 'Webinar',
 			modalidade: 'Presencial',
 			sincronismo: 'Ao vivo',
+			canal: 'Aberto',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			atualizarInscritos: true,
+			local: '',
 			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
 			notificarResponsavel: true,
+			rotuloContato: 'Mande-nos um e-mail',
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: novasCategorias,
-			situacao: 'Em desenvolvimento'
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			mensagemAnexo: fakerPT_BR.lorem.sentence(),
+			visualizacao: 'Colaborador',
+			situacao: 'Em desenvolvimento',
+			notificarConcluirPrimeiraAula: 'Não',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: false,
+			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
+			habilitarPagamento: false
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -299,7 +386,7 @@ describe('trilha', () => {
 		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
 
 		const todasCategorias = [...categorias, ...novasCategorias]
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
 
 		// DELETE
@@ -309,8 +396,8 @@ describe('trilha', () => {
 		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
 	})
 
-	it('4. CRUD trilha suspensa, com confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
+	it('4. CRUD catalogo suspenso, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
+		// Massa de dados para criação do catálogo
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
@@ -318,44 +405,52 @@ describe('trilha', () => {
 			horaInicio: '01:00',
 			dataFim: gerarData(),
 			horaFim: '23:00',
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
 			tipo: 'Outros',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
+			canal: 'Aberto',
 			cargaHoraria: fakerPT_BR.number.int({ min: 100, max: 999 }),
 			vigencia: fakerPT_BR.number.int({ min: 100, max: 999 }),
 			local: 'Youtube',
 			emailResponsavel: fakerPT_BR.internet.email(),
-			notificarResponsavel: false,
+			site: fakerPT_BR.internet.url(),
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
+			visualizacao: 'Colaborador',
 			situacao: 'Suspenso',
+			exigeConfirmacao: 'Habilitado',
+			diasTeste: fakerPT_BR.number.int({ min: 100, max: 999 }),
 			exigeConfirmacao: 'Habilitado'
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-        cy.validarDadosConteudo(dadosParaValidar, categorias)
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// UPDATE
 		cy.log('## UPDATE ##')
-
+		
 		novasCategorias = [`Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
 		const conteudoEdit = {
 			dataInicio: '01/01/2000',
 			dataFim: '28/02/2030',
 			tipo: 'Treinamento',
+			canal: 'Em companhia',
 			addCategoria: novasCategorias,
+			visualizacao: 'Usuários',
 			situacao: 'Em desenvolvimento',
 			exigeConfirmacao: 'Desabilitado'
 		}
@@ -369,36 +464,44 @@ describe('trilha', () => {
 		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		const todasCategorias = [...categorias, ...novasCategorias]
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
 
 		// DELETE
 		cy.log('## DELETE ##')
-		
+
 		cy.cancelarFormularioConteudo(tipoConteudo)
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('5. CRUD trilha suspensa, sem confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
+	it('5. CRUD catalogo em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
+		// Massa de dados para criação do catálogo
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
 			tipo: 'Palestra',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
+			canal: 'Aberto',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1000, max: 9999 }),
 			vigencia: fakerPT_BR.number.int({ min: 1000, max: 9999 }),
 			local: 'Twygo',
 			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
+			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
-			situacao: 'Suspenso'
+			visualizacao: 'Usuários',
+			situacao: 'Suspenso',
+			notificarConcluirPrimeiraAula: 'Não',
+			diasTeste: fakerPT_BR.number.int({ min: 1000, max: 9999 }),
+			exigeConfirmacao: 'Habilitado'
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
+		cy.acessarPgCatalogo()	
 		cy.addConteudo(tipoConteudo)
 		cy.preencherDadosConteudo(conteudo, { limpar: true })
 		cy.salvarConteudo(conteudo.nome, tipoConteudo)
@@ -410,12 +513,13 @@ describe('trilha', () => {
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
-		
+
 		// UPDATE
 		cy.log('## UPDATE ##')
-
+		
 		const conteudoEdit = {
-			tipo: 'Outros'
+			tipo: 'Outros',
+			visualizacao: 'Inscritos'
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -426,7 +530,7 @@ describe('trilha', () => {
 
 		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// DELETE
@@ -436,51 +540,49 @@ describe('trilha', () => {
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('6. CRUD trilha em desenvolvimento, sem confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
+	it('6. CRUD catalogo liberado, sem anexo, sem pagamento, sem confirmação, com visualização para público', () => {
+		// Massa de dados para criação do catálogo
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
 			tipo: 'Webinar',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
+			canal: 'Aberto',
 			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 99 }),
 			vigencia: fakerPT_BR.number.int({ min: 1, max: 99 }),
 			local: 'Zoom',
 			emailResponsavel: fakerPT_BR.internet.email(),
-			notificarResponsavel: false,
+			site: fakerPT_BR.internet.url(),
 			addCategoria: categorias,
-			situacao: 'Em desenvolvimento'
+			visualizacao: 'Público',
+			situacao: 'Liberado',
+			notificarConcluirPrimeiraAula: 'Sim',
+			notificarUsuarios: 'Sim'
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-        cy.validarDadosConteudo(dadosParaValidar, categorias)
-		
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
+
 		// UPDATE
 		cy.log('## UPDATE ##')
-
+		
 		const conteudoEdit = {
-			tipo: 'Palestra',
-			modalidade: 'Presencial',
-			cep: '85803-760',
-			endereco: 'Av. das Torres',
-			complemento: 'Apto 202',
-			cidade: 'Cascavel',
-			estado: 'PR',
-			pais: 'Brasil'
+			tipo: 'Palestra'
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -491,7 +593,7 @@ describe('trilha', () => {
 
 		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// DELETE
@@ -501,41 +603,52 @@ describe('trilha', () => {
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('7. CRUD trilha em desenvolvimento, com confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação da trilha
+	it('7. CRUD catalogo em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para usuários', () => {
+		// Massa de dados para criação do catálogo
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			descricao: fakerPT_BR.commerce.productDescription(),
+			tipo: 'Treinamento',
+			modalidade: 'Online',
 			sincronismo: 'Ao vivo',
+			canal: 'Aberto',
 			vigencia: '10000',
 			local: 'Twitch',
-			addCategoria: categorias
+			site: fakerPT_BR.internet.url(),
+			hashtag: `#${fakerPT_BR.hacker.abbreviation()}`,
+			addCategoria: categorias,
+			visualizacao: 'Usuários',
+			situacao: 'Em desenvolvimento',
+			notificarConcluirPrimeiraAula: 'Não'
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-        cy.addConteudo(tipoConteudo)
-        cy.preencherDadosConteudo(conteudo, { limpar: true })
-        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+		cy.acessarPgCatalogo()	
+		cy.addConteudo(tipoConteudo)
+		cy.preencherDadosConteudo(conteudo, { limpar: true })
+		cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-        cy.editarConteudo(conteudo.nome, tipoConteudo)
+		cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-        cy.validarDadosConteudo(dadosParaValidar, categorias)		
+		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// UPDATE
 		cy.log('## UPDATE ##')
-
-		delCategorias = categorias[0]
+		
+		delCategorias = [categorias[0], categorias[1]]
 		const conteudoEdit = {
-			vigencia: '0',
-			atualizarInscritos: true,
-			removerCategoria: delCategorias
+			vigencia: '',
+			site: '',
+			hashtag: '',
+			removerCategoria: delCategorias[0],
+			removerCategoria: delCategorias[1]
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -545,12 +658,12 @@ describe('trilha', () => {
 		cy.log('## READ-UPDATE ##')
 
 		cy.editarConteudo(conteudo.nome, tipoConteudo)
-		
+
 		const todasCategorias = categorias.filter(categoria => 
 			!delCategorias.includes(categoria)
 		)
-		
-		dadosParaValidar = { ...formularioConteudo, ...conteudo, ...conteudoEdit }
+
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
 
 		// DELETE
