+++ cypress/e2e/configuracoes/cobrancaDeInscricao/cobrancaAutomatica.cy.js
