+++ cypress/e2e/conteudos/curso/curso.cy.js
@@ -1,14 +1,13 @@
 /// reference types="cypress" />
 import 'cypress-real-events/support'
 import { fakerPT_BR } from '@faker-js/faker'
-import { getAuthToken } from '../support/authHelper'
-import { gerarData } from '../support/utilsHelper'
-import formConteudos from "../support/pageObjects/formConteudos"
+import { getAuthToken } from '../../../support/authHelper'
+import { gerarData } from '../../../support/utilsHelper'
 
-describe('catálogo', () => {
+describe('curso', () => {
 	let nome, tipoConteudo, categorias, novasCategorias, delCategorias
 
-	// Campos e dados default do formulário de catálogo
+	// Campos e dados default do formulário de curso
 	let formularioConteudo = {
 		nome: '',
 		dataInicio: '',
@@ -51,11 +50,12 @@ describe('catálogo', () => {
 		habilitarPagamento: false,
 		nrParcelas: '1',
 		valorAcrescimo: '0.0',
+		habilitarChat: false
 	}
 
-	beforeEach(() => {
+	beforeEach( () => {
 		// Define o tipo de conteúdo
-		tipoConteudo = 'catalogo'
+		tipoConteudo = 'curso'
 
 		// Gera um nome aleatório para o conteúdo
 		nome = fakerPT_BR.commerce.productName()
@@ -64,34 +64,33 @@ describe('catálogo', () => {
 		categorias = []
 		novasCategorias = []
 		delCategorias = []
-		
-		// Obtem o token de autenticação
+
+		// Obtém o token de autenticação
 		getAuthToken()
 
-		// Exclui todos os catálogos antes de iniciar o teste
-		cy.excluirCatalogoViaApi()
+		// Exclui todos os cursos antes de iniciar o teste
+		cy.excluirCursoViaApi()
 	})
-	
-	it('1. CRUD catalogo com dados default', () => {
-		// Massa de dados para criação do catálogo
-		const conteudo = {
-			nome: fakerPT_BR.commerce.productName(),
-			descricao: fakerPT_BR.commerce.productDescription()
+
+	it('1. CRUD curso com dados default', () =>{
+		// Massa de dados para criação do curso
+        const conteudo = {
+			nome: nome,
+            descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
-
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
+        
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
 
@@ -106,7 +105,7 @@ describe('catálogo', () => {
 			horaInicio: '12:00',
 			dataFim: '29/04/2024',
 			horaFim: '22:00',
-			descricao: `Descrição editada do curso nome: ${novoNome}`,
+			descricao: `${fakerPT_BR.commerce.productDescription()} editado do evento ${novoNome}`,
 			tipo: 'Congresso',
 			modalidade: 'Presencial',
 			sincronismo: 'Ao vivo',
@@ -128,8 +127,8 @@ describe('catálogo', () => {
 			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
 			permiteAnexo: 'Habilitado',
+			mensagemAnexo: `Insira o anexo do Curso: ${novoNome}`,
 			statusIframeAnexo: true,
-			mensagemAnexo: 'Insira o anexo do Catálogo do evento:',
 			visualizacao: 'Público',
 			situacao: 'Liberado',
 			notificarConcluirPrimeiraAula: 'Sim',
@@ -137,244 +136,11 @@ describe('catálogo', () => {
 			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
 			habilitarDiasTeste: true,
 			exigeConfirmacao: 'Habilitado',
-			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
+			valorInscricao: fakerPT_BR.commerce.price({ min: 60, max: 100 }),
 			habilitarPagamento: true,
 			nrParcelas: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 })
-		}
-		
-		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
-		cy.salvarConteudo(conteudoEdit.nome, tipoConteudo)
-
-		// READ-UPDATE
-		cy.log('## READ-UPDATE ##')
-
-		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
-
-		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
-		// DELETE
-		cy.log('## DELETE ##')
-
-		cy.cancelarFormularioConteudo(tipoConteudo)
-		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
-	})
-
-	it('2. CRUD catalogo liberado, com anexo, com pagamento, sem acréscimo, com confirmação, com visualização para inscritos', () => {
-		// Massa de dados para criação do catálogo
-		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
-		const conteudo = {
-			nome: nome,
-			dataInicio: '29/03/2024',
-			horaInicio: '01:00',
-			dataFim: '29/04/2024',
-			horaFim: '23:00',
-			descricao: fakerPT_BR.commerce.productDescription(),
-			tipo: 'Congresso',
-			modalidade: 'Presencial',
-			sincronismo: 'Ao vivo',
-			canal: 'Outros',
-			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			local: 'Centro de Eventos',
-			cep: '85804-455',
-			endereco: 'Rua das Flores',
-			complemento: 'Casa dos fundos',
-			cidade: 'Joinville',
-			estado: 'SC',
-			pais: 'Brasil',
-			emailResponsavel: fakerPT_BR.internet.email(),
-			site: fakerPT_BR.internet.url(),
-			notificarResponsavel: true,
-			rotuloContato: 'Fale conosco',
-			hashtag: fakerPT_BR.hacker.adjective(),
-			addCategoria: categorias,
-			permiteAnexo: 'Habilitado',
-			statusIframeAnexo: true,
-			mensagemAnexo: `Insira o anexo do Catálogo do evento: ${nome}`,
-			visualizacao: 'Inscritos',
-			situacao: 'Liberado',
-			notificarConcluirPrimeiraAula: 'Sim',
-			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			habilitarDiasTeste: true,
-			exigeConfirmacao: 'Habilitado',
-			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
-			habilitarPagamento: true,
-			nrParcelas: fakerPT_BR.number.int({ min: 1, max: 9 })
-		}
-
-		// CREATE
-		cy.log('## CREATE ##')
-
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
-
-		// READ
-		cy.log('## READ ##')
-
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
-
-		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
-		// UPDATE
-		cy.log('## UPDATE ##')
-
-		const novoNome = fakerPT_BR.commerce.productName()
-		categorias = [`Cat3-${fakerPT_BR.hacker.noun()}`]
-		const conteudoEdit = {
-			nome: novoNome,
-			dataInicio: '01/01/2023',
-			horaInicio: '00:01',
-			dataFim: '31/01/2025',
-			horaFim: '23:59',
-			descricao: `Descrição editada do curso nome: ${novoNome}`,
-			tipo: 'Feira',
-			modalidade: 'Online',
-			sincronismo: 'Gravado',
-			canal: '',
-			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			atualizarInscritos: true,
-			local: 'T&D Connect',
-			emailResponsavel: fakerPT_BR.internet.email(),
-			site: fakerPT_BR.internet.url(),
-			notificarResponsavel: false,
-			rotuloContato: 'Contato',
-			hashtag: fakerPT_BR.hacker.adjective(),
-			addCategoria: categorias,
-			permiteAnexo: 'Desabilitado',
-			statusIframeAnexo: false,
-			visualizacao: 'Público',
-			situacao: 'Suspenso',
-			notificarConcluirPrimeiraAula: 'Não',
-			notificarUsuarios: 'Sim',
-			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			habilitarDiasTeste: false,
-			exigeConfirmacao: 'Desabilitado',
-			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
-			habilitarPagamento: false
-		}
-
-		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
-		cy.salvarConteudo(conteudoEdit.nome, tipoConteudo)
-
-		// READ-UPDATE
-		cy.log('## READ-UPDATE ##')
-
-		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
-
-		// Após salvar a alteração de modalidade de Presencial para Online, o campo de CEP não é limpo como os demais campos de endereço
-		let dadosEspecificos = {
-			cep: '85804-455',
-			endereco: '',
-			complemento: '',
-			cidade: '',
-			estado: '',
-			pais: ''
-		}
-
-		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit, ...dadosEspecificos }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
-		// DELETE
-		cy.log('## DELETE ##')
-
-		cy.cancelarFormularioConteudo(tipoConteudo)
-		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
-	})
-
-	it('3. CRUD catalogo liberado, com anexo, com pagamento, c/acréscimo, sem confirmação, com visualização para público', () => {
-		// Massa de dados para criação do catálogo
-		const categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
-		const conteudo = {
-			nome: nome,
-			dataInicio: '29/01/2023',
-			horaInicio: '11:20',
-			dataFim: '29/01/2024',
-			horaFim: '22:02',
-			descricao: fakerPT_BR.commerce.productDescription(),
-			tipo: 'Feira',
-			modalidade: 'Online',
-			sincronismo: 'Gravado',
-			canal: 'Em companhia',
-			cargaHoraria: fakerPT_BR.number.int({ min: 10, max: 99 }),
-			vigencia: fakerPT_BR.number.int({ min: 10, max: 99 }),
-			local: 'Youtube',
-			emailResponsavel: fakerPT_BR.internet.email(),
-			site: fakerPT_BR.internet.url(),
-			rotuloContato: 'Entre em contato conosco para mais informações.',
-			hashtag: fakerPT_BR.hacker.adjective(),
-			addCategoria: categorias,
-			permiteAnexo: 'Habilitado',
-			statusIframeAnexo: true,
-			visualizacao: 'Público',
-			situacao: 'Liberado',
-			notificarConcluirPrimeiraAula: 'Sim',
-			notificarUsuarios: 'Sim',
-			diasTeste: fakerPT_BR.number.int({ min: 10, max: 99 }),
-			habilitarDiasTeste: true,
-			valorInscricao: fakerPT_BR.commerce.price({ min: 50, max: 99 }),
-			habilitarPagamento: true,
-			nrParcelas: fakerPT_BR.number.int({ min: 10, max: 12 }),
-			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 })
-		}
-
-		// CREATE
-		cy.log('## CREATE ##')
-
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
-
-		// READ
-		cy.log('## READ ##')
-
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
-
-		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
-		// UPDATE
-		cy.log('## UPDATE ##')
-
-		const novoNome = fakerPT_BR.commerce.productName()
-		novasCategorias = [`Cat4-${fakerPT_BR.hacker.noun()}`, `Cat5-${fakerPT_BR.hacker.noun()}`]
-		const conteudoEdit = {
-			nome: novoNome,
-			dataInicio: '10/03/2000',
-			horaInicio: '00:00',
-			dataFim: '31/12/2050',
-			horaFim: '03:40',
-			descricao: `Descrição editada do curso nome: ${novoNome}`,
-			tipo: 'Webinar',
-			modalidade: 'Presencial',
-			sincronismo: 'Ao vivo',
-			canal: 'Aberto',
-			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			atualizarInscritos: true,
-			local: '',
-			emailResponsavel: fakerPT_BR.internet.email(),
-			site: fakerPT_BR.internet.url(),
-			notificarResponsavel: true,
-			rotuloContato: 'Mande-nos um e-mail',
-			hashtag: fakerPT_BR.hacker.adjective(),
-			addCategoria: novasCategorias,
-			permiteAnexo: 'Habilitado',
-			statusIframeAnexo: true,
-			mensagemAnexo: fakerPT_BR.lorem.sentence(),
-			visualizacao: 'Colaborador',
-			situacao: 'Em desenvolvimento',
-			notificarConcluirPrimeiraAula: 'Não',
-			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
-			habilitarDiasTeste: false,
-			valorInscricao: fakerPT_BR.commerce.price({ min: 45, max: 90 }),
-			habilitarPagamento: false
+			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 }),
+			habilitarChat: true
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -396,8 +162,245 @@ describe('catálogo', () => {
 		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
 	})
 
-	it('4. CRUD catalogo suspenso, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
-		// Massa de dados para criação do catálogo
+	it('2. CRUD curso liberado, com anexo, com pagamento, sem acréscimo, com confirmação, com visualização para inscritos', () => {
+		// Massa de dados para criação do curso
+		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
+		const conteudo = {
+			nome: nome,
+			dataInicio: '29/03/2024',
+			horaInicio: '01:00',
+			dataFim: '29/04/2024',
+			horaFim: '23:00',
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			tipo: 'Congresso',
+			modalidade: 'Presencial',
+			sincronismo: 'Ao vivo',
+			canal: 'Outros',
+			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			local: 'Centro de Eventos',
+			cep: '85804-455',
+			endereco: 'Rua das Flores',
+			complemento: 'Casa dos fundos',
+			cidade: 'Joinville',
+			estado: 'SC',
+			pais: 'Brasil',
+			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: true,
+			rotuloContato: 'Fale conosco',
+			hashtag: fakerPT_BR.hacker.adjective(),
+			addCategoria: categorias,
+			permiteAnexo: 'Habilitado',
+			mensagemAnexo: `Insira o anexo do Curso: ${nome}`,
+			statusIframeAnexo: true,
+			visualizacao: 'Inscritos',
+			situacao: 'Liberado',
+			notificarConcluirPrimeiraAula: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: true,
+			exigeConfirmacao: 'Habilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 60, max: 100 }),
+			habilitarPagamento: true,
+			nrParcelas: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarChat: true
+		}
+
+		// CREATE
+		cy.log('## CREATE ##')
+
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+
+		// READ
+		cy.log('## READ ##')
+
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
+
+		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
+        cy.validarDadosConteudo(dadosParaValidar, categorias)
+
+		// UPDATE
+		cy.log('## UPDATE ##')
+
+		const novoNome = fakerPT_BR.commerce.productName()
+		novasCategorias = [`Cat3-${fakerPT_BR.hacker.noun()}`, `Cat4-${fakerPT_BR.hacker.noun()}`]
+		const conteudoEdit = {
+			nome: novoNome,
+			dataInicio: '01/01/2023',
+			horaInicio: '00:01',
+			dataFim: '31/01/2025',
+			horaFim: '23:59',
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${novoNome}`,
+			tipo: 'Feira',
+			modalidade: 'Online',
+			sincronismo: 'Gravado',
+			canal: '',
+			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			atualizarInscritos: true,
+			local: 'T&D Connect',
+			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: false,
+			rotuloContato: 'Contato',
+			hashtag: fakerPT_BR.hacker.adjective(),
+			addCategoria: novasCategorias,
+			permiteAnexo: 'Desabilitado',
+			statusIframeAnexo: false,
+			visualizacao: 'Público',
+			situacao: 'Suspenso',
+			notificarConcluirPrimeiraAula: 'Não',
+			notificarUsuarios: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: false,
+			exigeConfirmacao: 'Desabilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 5, max: 9 }),
+			habilitarPagamento: false,
+			habilitarChat: false
+		}
+
+		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
+		cy.salvarConteudo(conteudoEdit.nome, tipoConteudo)
+
+		// READ-UPDATE
+		cy.log('## READ-UPDATE ##')
+
+		let dadosEspecificos = {
+			endereco: '',
+			complemento: '',
+			cidade: '',
+			estado: '',
+			pais: ''
+		}
+
+		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
+
+		const todasCategorias = [...categorias, ...novasCategorias]
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit, ...dadosEspecificos }		
+		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
+
+		// DELETE
+		cy.log('## DELETE ##')
+		
+		cy.cancelarFormularioConteudo(tipoConteudo)
+		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
+	})
+
+	it('3. CRUD curso liberado, com anexo, com pagamento, c/acréscimo, sem confirmação, com visualização para público', () => {
+		// Massa de dados para criação do curso
+		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
+		const conteudo = {
+			nome: nome,
+			dataInicio: '29/01/2023',
+			horaInicio: '11:20',
+			dataFim: '29/01/2024',
+			horaFim: '22:02',
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
+			tipo: 'Feira',
+			modalidade: 'Online',
+			sincronismo: 'Gravado',
+			canal: 'Em companhia',
+			cargaHoraria: fakerPT_BR.number.int({ min: 10, max: 99 }),
+			vigencia: fakerPT_BR.number.int({ min: 10, max: 99 }),
+			local: 'Youtube',
+			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: false,
+			rotuloContato: 'Entre em contato conosco para mais informações.',
+			hashtag: fakerPT_BR.hacker.adjective(),
+			addCategoria: categorias,
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			visualizacao: 'Público',
+			situacao: 'Liberado',
+			notificarConcluirPrimeiraAula: 'Sim',
+			notificarUsuarios: 'Sim',
+			diasTeste: fakerPT_BR.number.int({ min: 10, max: 99 }),
+			habilitarDiasTeste: true,
+			valorInscricao: fakerPT_BR.commerce.price({ min: 100, max: 999 }),
+			habilitarPagamento: true,
+			nrParcelas: fakerPT_BR.number.int({ min: 10, max: 12 }),
+			valorAcrescimo: fakerPT_BR.commerce.price({ min: 1, max: 9, dec: 1 }),
+			habilitarChat: true
+		}
+
+		// CREATE
+		cy.log('## CREATE ##')
+
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
+
+		// READ
+		cy.log('## READ ##')
+
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
+
+		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
+        cy.validarDadosConteudo(dadosParaValidar, categorias)
+
+		// UPDATE
+		cy.log('## UPDATE ##')
+
+		const novoNome = fakerPT_BR.commerce.productName()
+		novasCategorias = [`Cat4-${fakerPT_BR.hacker.noun()}`, `Cat5-${fakerPT_BR.hacker.noun()}`]		
+		const conteudoEdit = {
+			nome: novoNome,
+			dataInicio: '10/03/2000',
+			horaInicio: '00:00',
+			dataFim: '31/12/2050',
+			horaFim: '03:40',
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${novoNome}`,
+			tipo: 'Webinar',
+			modalidade: 'Presencial',
+			sincronismo: 'Ao vivo',
+			canal: 'Aberto',
+			cargaHoraria: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			vigencia: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			atualizarInscritos: true,
+			emailResponsavel: fakerPT_BR.internet.email(),
+			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: true,
+			rotuloContato: 'Mande-nos um e-mail',
+			hashtag: fakerPT_BR.hacker.adjective(),
+			addCategoria: novasCategorias,
+			permiteAnexo: 'Habilitado',
+			statusIframeAnexo: true,
+			mensagemAnexo: `Insira o anexo do Curso: ${novoNome}`,
+			visualizacao: 'Colaborador',
+			situacao: 'Em desenvolvimento',
+			notificarConcluirPrimeiraAula: 'Não',
+			diasTeste: fakerPT_BR.number.int({ min: 1, max: 9 }),
+			habilitarDiasTeste: false,
+			exigeConfirmacao: 'Desabilitado',
+			valorInscricao: fakerPT_BR.commerce.price({ min: 5, max: 9 }),
+			habilitarPagamento: false,
+			habilitarChat: false
+		}
+
+		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
+		cy.salvarConteudo(conteudoEdit.nome, tipoConteudo)
+
+		// READ-UPDATE
+		cy.log('## READ-UPDATE ##')
+
+		cy.editarConteudo(conteudoEdit.nome, tipoConteudo)
+
+		const todasCategorias = [...categorias, ...novasCategorias]
+		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
+		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
+
+		// DELETE
+		cy.log('## DELETE ##')
+
+		cy.cancelarFormularioConteudo(tipoConteudo)
+		cy.excluirConteudo(conteudoEdit.nome, tipoConteudo)
+	})
+
+	it('4. CRUD curso suspenso, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
+		// Massa de dados para criação do curso
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
@@ -405,7 +408,7 @@ describe('catálogo', () => {
 			horaInicio: '01:00',
 			dataFim: gerarData(),
 			horaFim: '23:00',
-			descricao: fakerPT_BR.commerce.productDescription(),
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
 			tipo: 'Outros',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
@@ -415,34 +418,39 @@ describe('catálogo', () => {
 			local: 'Youtube',
 			emailResponsavel: fakerPT_BR.internet.email(),
 			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: false,
 			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
+			permiteAnexo: 'Desabilitado',
+			statusIframeAnexo: false,
 			visualizacao: 'Colaborador',
 			situacao: 'Suspenso',
-			exigeConfirmacao: 'Habilitado',
+			notificarConcluirPrimeiraAula: 'Não',
 			diasTeste: fakerPT_BR.number.int({ min: 100, max: 999 }),
-			exigeConfirmacao: 'Habilitado'
+			habilitarDiasTeste: false,
+			exigeConfirmacao: 'Habilitado',
+			habilitarPagamento: false,
+			habilitarChat: true
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
+        cy.validarDadosConteudo(dadosParaValidar, categorias)
 
 		// UPDATE
 		cy.log('## UPDATE ##')
-		
+
 		novasCategorias = [`Cat2-${fakerPT_BR.hacker.noun()}`, `Cat3-${fakerPT_BR.hacker.noun()}`]
 		const conteudoEdit = {
 			dataInicio: '01/01/2000',
@@ -469,17 +477,17 @@ describe('catálogo', () => {
 
 		// DELETE
 		cy.log('## DELETE ##')
-
+		
 		cy.cancelarFormularioConteudo(tipoConteudo)
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('5. CRUD catalogo em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
-		// Massa de dados para criação do catálogo
+	it('5. CRUD curso em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para colaboradores', () => {
+		// Massa de dados para criação do curso
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: fakerPT_BR.commerce.productDescription(),
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
 			tipo: 'Palestra',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
@@ -491,17 +499,18 @@ describe('catálogo', () => {
 			site: fakerPT_BR.internet.url(),
 			hashtag: fakerPT_BR.hacker.adjective(),
 			addCategoria: categorias,
+			statusIframeAnexo: false,
 			visualizacao: 'Usuários',
 			situacao: 'Suspenso',
 			notificarConcluirPrimeiraAula: 'Não',
 			diasTeste: fakerPT_BR.number.int({ min: 1000, max: 9999 }),
-			exigeConfirmacao: 'Habilitado'
+			exigeConfirmacao: 'Habilitado',
+			habilitarChat: true
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()	
 		cy.addConteudo(tipoConteudo)
 		cy.preencherDadosConteudo(conteudo, { limpar: true })
 		cy.salvarConteudo(conteudo.nome, tipoConteudo)
@@ -513,13 +522,14 @@ describe('catálogo', () => {
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
 		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
+		
 		// UPDATE
 		cy.log('## UPDATE ##')
-		
+
 		const conteudoEdit = {
 			tipo: 'Outros',
-			visualizacao: 'Inscritos'
+			visualizacao: 'Inscritos',
+			statusIframeAnexo: false
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -540,12 +550,12 @@ describe('catálogo', () => {
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('6. CRUD catalogo liberado, sem anexo, sem pagamento, sem confirmação, com visualização para público', () => {
-		// Massa de dados para criação do catálogo
+	it('6. CRUD curso liberado, sem anexo, sem pagamento, sem confirmação, com visualização para público', () => {
+		// Massa de dados para criação do curso
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: fakerPT_BR.commerce.productDescription(),
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
 			tipo: 'Webinar',
 			modalidade: 'Online',
 			sincronismo: 'Gravado',
@@ -555,32 +565,36 @@ describe('catálogo', () => {
 			local: 'Zoom',
 			emailResponsavel: fakerPT_BR.internet.email(),
 			site: fakerPT_BR.internet.url(),
+			notificarResponsavel: false,
 			addCategoria: categorias,
+			statusIframeAnexo: false,
 			visualizacao: 'Público',
 			situacao: 'Liberado',
 			notificarConcluirPrimeiraAula: 'Sim',
-			notificarUsuarios: 'Sim'
+			habilitarDiasTeste: false,
+			notificarUsuarios: 'Sim',
+			habilitarPagamento: false,
+			habilitarChat: true
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
-
+        cy.validarDadosConteudo(dadosParaValidar, categorias)
+		
 		// UPDATE
 		cy.log('## UPDATE ##')
-		
+
 		const conteudoEdit = {
 			tipo: 'Palestra'
 		}
@@ -603,14 +617,12 @@ describe('catálogo', () => {
 		cy.excluirConteudo(conteudo.nome, tipoConteudo)
 	})
 
-	it('7. CRUD catalogo em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para usuários', () => {
-		// Massa de dados para criação do catálogo
+	it('7. CRUD curso em desenvolvimento, sem anexo, sem pagamento, com confirmação, com visualização para usuários', () => {
+		// Massa de dados para criação do curso
 		categorias = [`Cat1-${fakerPT_BR.hacker.noun()}`, `Cat2-${fakerPT_BR.hacker.noun()}`]
 		const conteudo = {
 			nome: nome,
-			descricao: fakerPT_BR.commerce.productDescription(),
-			tipo: 'Treinamento',
-			modalidade: 'Online',
+			descricao: `${fakerPT_BR.commerce.productDescription()} do evento ${nome}`,
 			sincronismo: 'Ao vivo',
 			canal: 'Aberto',
 			vigencia: '10000',
@@ -619,36 +631,33 @@ describe('catálogo', () => {
 			hashtag: `#${fakerPT_BR.hacker.abbreviation()}`,
 			addCategoria: categorias,
 			visualizacao: 'Usuários',
-			situacao: 'Em desenvolvimento',
-			notificarConcluirPrimeiraAula: 'Não'
+			habilitarChat: true
 		}
 
 		// CREATE
 		cy.log('## CREATE ##')
 
-		cy.acessarPgCatalogo()	
-		cy.addConteudo(tipoConteudo)
-		cy.preencherDadosConteudo(conteudo, { limpar: true })
-		cy.salvarConteudo(conteudo.nome, tipoConteudo)
+        cy.addConteudo(tipoConteudo)
+        cy.preencherDadosConteudo(conteudo, { limpar: true })
+        cy.salvarConteudo(conteudo.nome, tipoConteudo)
 
 		// READ
 		cy.log('## READ ##')
 
-		cy.editarConteudo(conteudo.nome, tipoConteudo)
+        cy.editarConteudo(conteudo.nome, tipoConteudo)
 
 		let dadosParaValidar = { ...formularioConteudo, ...conteudo }
-		cy.validarDadosConteudo(dadosParaValidar, categorias)
+        cy.validarDadosConteudo(dadosParaValidar, categorias)		
 
 		// UPDATE
 		cy.log('## UPDATE ##')
-		
-		delCategorias = [categorias[0], categorias[1]]
+
+		delCategorias = categorias[0]
 		const conteudoEdit = {
-			vigencia: '',
-			site: '',
-			hashtag: '',
-			removerCategoria: delCategorias[0],
-			removerCategoria: delCategorias[1]
+			vigencia: '0',
+			atualizarInscritos: true,
+			removerCategoria: delCategorias,
+			habilitarChat: false
 		}
 
 		cy.preencherDadosConteudo(conteudoEdit, { limpar: true })
@@ -658,11 +667,11 @@ describe('catálogo', () => {
 		cy.log('## READ-UPDATE ##')
 
 		cy.editarConteudo(conteudo.nome, tipoConteudo)
-
+		
 		const todasCategorias = categorias.filter(categoria => 
 			!delCategorias.includes(categoria)
 		)
-
+		
 		dadosParaValidar = { ...dadosParaValidar, ...conteudoEdit }
 		cy.validarDadosConteudo(dadosParaValidar, todasCategorias)
