+++ cypress/e2e/conteudos/curso/cursoAmbienteAdicional.cy.js
