+++ cypress/e2e/usuarios/usuario.cy.js
@@ -1,6 +1,6 @@
 /// reference types="cypress" />
 import { fakerPT_BR } from '@faker-js/faker'
-import { getAuthToken } from '../support/authHelper'
+import { getAuthToken } from '../../support/authHelper'
 let faker = require('faker-br')
 
 describe('Usuário', () => {
